Wrap newM3Connection with Form
