+++ Source/SelfService/Web/integrations/connection/new/components/MetadataPublisherCredentials.tsx
@@ -8,10 +8,10 @@ import { Stack } from '@mui/material';
 import { Form, Input } from '@dolittle/design-system';
 import { MaxWidthTextBlock } from './MaxWidthTextBlock';
 
-type MetadataPublisherParameters = {
-    metadataPublisher: string;
-    password: string;
-};
+// type MetadataPublisherParameters = {
+//     metadataPublisher: string;
+//     password: string;
+// };
 
 export const MetadataPublisherCredentials = () =>
     <>
@@ -19,23 +19,14 @@ export const MetadataPublisherCredentials = () =>
             This will allow us to access your service and provide the data, including custom data fields, needed to configure your application logic.
         </MaxWidthTextBlock>
 
-        <Form<MetadataPublisherParameters>
-            initialValues={{
-                metadataPublisher: '',
-                password: '',
-            }}
-            onSubmit={() => { }}
-            sx={{ mt: 3 }}
-        >
-            <Stack spacing={3.5}>
-                <Input
-                    id='metadataPublisher'
-                    label='Your Metadata Publisher URL'
-                    placeholder='Your metadata publisher URL goes here...'
-                    required
-                    sx={{ width: 1, maxWidth: 500 }}
-                />
-                <Input id='password' label='Password' placeholder='Your password' required />
-            </Stack>
-        </Form>
+        <Stack spacing={3.5} sx={{ mt: 3 }}>
+            <Input
+                id='metadataPublisher'
+                label='Your Metadata Publisher URL'
+                placeholder='Your metadata publisher URL goes here...'
+                required
+                sx={{ width: 1, maxWidth: 500 }}
+            />
+            <Input id='metadataPublisherPassword' label='Password' placeholder='Your password' required />
+        </Stack>
     </>;
